requirements.txt
# Binance Futures Trading Bot Dependencies


# Core Dependencies
python-binance==1.0.15

# Optional: For enhanced logging and formatting
#colorama==0.4.6